@@ -9,7 +9,7 @@
     "dist": "npm run clean && npm run copy && cross-env NODE_ENV=production webpack --progress --config build/webpack.config.js",
     "lint": "eslint --max-warnings 0 ./src",
     "posttest": "npm run lint",
-    "start": "cross-env NODE_ENV=development webpack-dev-server --progress --hot --config build/webpack.config.js",
+    "start": "cross-env NODE_ENV=development webpack-dev-server --progress --config build/webpack.config.js",
     "test": "istanbul cover karma start",
     "test:watch": "karma start --autoWatch=true --singleRun=false",
     "deploy": "git subtree push --prefix dist origin gh-pages",
@@ -37,12 +37,10 @@
     "babel-plugin-transform-async-to-generator": "^6.16.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-assign": "^6.8.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.2.10",
     "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.0.15",
     "babel-preset-react": "^6.0.15",
     "babel-preset-stage-0": "^6.5.0",
-    "bower-webpack-plugin": "^0.1.9",
     "chai": "^3.2.0",
     "ci-publish": "^1.3.0",
     "copyfiles": "^1.0.0",
@@ -54,11 +52,9 @@
     "eslint-plugin-react": "^6.0.0",
     "expect": "^1.20.2",
     "file-loader": "^0.9.0",
-    "friendly-errors-webpack-plugin": "^1.6.1",
     "glob": "^7.0.0",
     "html-minifier": "^3.1.0",
     "imports-loader": "^0.6.5",
-    "isparta-instrumenter-loader": "^1.0.0",
     "istanbul": "^0.4.5",
     "karma": "^1.0.0",
     "karma-chai": "^0.1.0",
@@ -76,26 +72,25 @@
     "open": "0.0.5",
     "open-browser-webpack-plugin": "^0.0.5",
     "react-addons-test-utils": "^15.0.0",
-    "react-hot-loader": "^1.3.0",
+    "react-hot-loader": "^v3.0.0-beta.6",
     "rimraf": "^2.4.3",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.6",
-    "webpack": "^1.13.3",
+    "webpack": "^2.x",
     "webpack-bundle-analyzer": "^1.5.3",
-    "webpack-dev-server": "^1.16.2",
+    "webpack-dev-server": "^2.x",
     "webpack-merge": "^4.1.0"
   },
   "dependencies": {
     "animation-frame": "^0.3.0",
     "classlist-polyfill": "^1.0.3",
     "classnames": "^2.2.5",
-    "core-js": "^2.0.0",
-    "es6-promise": "^4.0.5",
     "locutus": "^2.0.5",
     "material-design-icons": "^3.0.1",
     "material-design-lite": "^1.2.1",
     "moment": "^2.16.0",
+    "prop-types": "^15.5.8",
     "react": "^15.0.0",
     "react-addons-css-transition-group": "^15.4.0",
     "react-dom": "^15.0.0",
