@@ -1,6 +1,6 @@
-import config from 'config';
+import config from '@/config';
 
-import { cachedFetch } from '../utils';
+import { cachedFetch } from '@/utils';
 
 export const SET_ITEMS = 'SET_ITEMS';
 export const SET_IMAGES = 'SET_IMAGES';
