@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Message } from '../components';
+import { Message } from '@/components';
 
 export default class NotFoundContainer extends React.Component {
   constructor(props) {
