@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 
-import rootReducer from '../reducers';
+import rootReducer from '@/reducers';
 
 const createStoreWithMiddleware = applyMiddleware(thunkMiddleware)(createStore);
 
@@ -16,8 +16,8 @@ export default function configureStore(initialState) {
 
   if (module.hot) {
     // Enable Webpack hot module replacement for reducers
-    module.hot.accept('../reducers', () => {
-      const nextRootReducer = require('../reducers');
+    module.hot.accept('@/reducers', () => {
+      const nextRootReducer = require('@/reducers');
       store.replaceReducer(nextRootReducer);
     });
   }
