@@ -1,6 +1,6 @@
-export cachedFetch from './cachedFetch';
-export hashStr from './hashStr';
-export moment from './moment';
-export Piwik from './Piwik';
-export scrollTo from './scrollTo';
-export Storage from './Storage';
+export cachedFetch from '@/utils/cachedFetch';
+export hashStr from '@/utils/hashStr';
+export moment from '@/utils/moment';
+export Piwik from '@/utils/Piwik';
+export scrollTo from '@/utils/scrollTo';
+export Storage from '@/utils/Storage';
